+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Upcoming
+
+- Simplify babel-preset by importing `babel-preset-react`
+
 ## 1.4.1
 
 - Add support for `babel.plugins` in `.motion/config.json`